Cleanup ui components some
