+++ ui/packages/ui/package.json
@@ -20,6 +20,7 @@
     "build:styles": "tailwindcss -i ./src/styles/globals.css -o ./dist/index.css",
     "build:components": "tsc",
     "dev:styles": "tailwindcss -i ./src/styles/globals.css -o ./dist/index.css --watch",
+    "dev:styles": "tailwindcss -i ./src/styles/globals.css -o ./dist/index.css --watch",
     "dev:components": "tsc --watch"
   },
   "devDependencies": {
