+++ ui/packages/ui/src/components/nav-main.tsx
@@ -1,13 +1,11 @@
 import { IconCirclePlusFilled, IconMail, type Icon } from "@tabler/icons-react";
 
 import { Button } from "@repo/ui/components/ui/button";
-import {
-  SidebarGroup,
-  SidebarGroupContent,
-  SidebarMenu,
-  SidebarMenuButton,
-  SidebarMenuItem,
-} from "@repo/ui/components/ui/sidebar";
+import { SidebarGroup } from "./ui/sidebar/sidebar-group";
+import { SidebarGroupContent } from "./ui/sidebar/sidebar-group-content";
+import { SidebarMenu } from "./ui/sidebar/sidebar-menu";
+import { SidebarMenuItem } from "./ui/sidebar/sidebar-menu-item";
+import { SidebarMenuButton } from "./ui/sidebar/sidebar-menu-button";
 
 export function NavMain({
   items,
