+++ ui/packages/ui/src/components/ui/sidebar/sidebar-content.tsx
@@ -1,0 +1,18 @@
+import { mergeClassValues } from "../../../functions/merge-class-values";
+
+export function SidebarContent({
+  className,
+  ...props
+}: React.ComponentProps<"div">) {
+  return (
+    <div
+      data-slot="sidebar-content"
+      data-sidebar="content"
+      className={mergeClassValues(
+        "flex min-h-0 flex-1 flex-col gap-2 overflow-auto group-data-[collapsible=icon]:overflow-hidden",
+        className,
+      )}
+      {...props}
+    />
+  );
+}
