+++ ui/packages/ui/src/components/ui/sidebar/sidebar-inset.tsx
@@ -1,0 +1,18 @@
+import { mergeClassValues } from "../../../functions/merge-class-values";
+
+export function SidebarInset({
+  className,
+  ...props
+}: React.ComponentProps<"main">) {
+  return (
+    <main
+      data-slot="sidebar-inset"
+      className={mergeClassValues(
+        "bg-background relative flex w-full flex-1 flex-col",
+        "md:peer-data-[variant=inset]:m-2 md:peer-data-[variant=inset]:ml-0 md:peer-data-[variant=inset]:rounded-xl md:peer-data-[variant=inset]:shadow-sm md:peer-data-[variant=inset]:peer-data-[state=collapsed]:ml-2",
+        className,
+      )}
+      {...props}
+    />
+  );
+}
