+++ ui/packages/ui/src/components/ui/sidebar/sidebar-menu-skeleton.tsx
@@ -1,0 +1,46 @@
+"use client";
+
+import { Skeleton } from "@repo/ui/components/ui/skeleton";
+import { mergeClassValues } from "../../../functions/merge-class-values";
+import { useMemo } from "react";
+
+export function SidebarMenuSkeleton({
+  className,
+  showIcon = false,
+  ...props
+}: React.ComponentProps<"div"> & {
+  showIcon?: boolean;
+}) {
+  // Random width between 50 to 90%.
+  const width = useMemo(() => {
+    return `${Math.floor(Math.random() * 40) + 50}%`;
+  }, []);
+
+  return (
+    <div
+      data-slot="sidebar-menu-skeleton"
+      data-sidebar="menu-skeleton"
+      className={mergeClassValues(
+        "flex h-8 items-center gap-2 rounded-md px-2",
+        className,
+      )}
+      {...props}
+    >
+      {showIcon && (
+        <Skeleton
+          className="size-4 rounded-md"
+          data-sidebar="menu-skeleton-icon"
+        />
+      )}
+      <Skeleton
+        className="h-4 max-w-(--skeleton-width) flex-1"
+        data-sidebar="menu-skeleton-text"
+        style={
+          {
+            "--skeleton-width": width,
+          } as React.CSSProperties
+        }
+      />
+    </div>
+  );
+}
