+++ ui/packages/ui/src/components/ui/sidebar/sidebar-menu-sub.tsx
@@ -1,0 +1,19 @@
+import { mergeClassValues } from "../../../functions/merge-class-values";
+
+export function SidebarMenuSub({
+  className,
+  ...props
+}: React.ComponentProps<"ul">) {
+  return (
+    <ul
+      data-slot="sidebar-menu-sub"
+      data-sidebar="menu-sub"
+      className={mergeClassValues(
+        "border-sidebar-border mx-3.5 flex min-w-0 translate-x-px flex-col gap-1 border-l px-2.5 py-0.5",
+        "group-data-[collapsible=icon]:hidden",
+        className,
+      )}
+      {...props}
+    />
+  );
+}
