+++ ui/packages/ui/src/components/ui/sidebar/sidebar-trigger.tsx
@@ -1,0 +1,32 @@
+"use client";
+
+import { mergeClassValues } from "../../../functions/merge-class-values";
+import { Button } from "../button";
+import { useSidebar } from "./useSidebar";
+import { PanelLeftIcon } from "lucide-react";
+
+export function SidebarTrigger({
+  className,
+  onClick,
+  ...props
+}: React.ComponentProps<typeof Button>) {
+  const { toggleSidebar } = useSidebar();
+
+  return (
+    <Button
+      data-sidebar="trigger"
+      data-slot="sidebar-trigger"
+      variant="ghost"
+      size="icon"
+      className={mergeClassValues("size-7", className)}
+      onClick={(event) => {
+        onClick?.(event);
+        toggleSidebar();
+      }}
+      {...props}
+    >
+      <PanelLeftIcon />
+      <span className="sr-only">Toggle Sidebar</span>
+    </Button>
+  );
+}
